move setup into init command, could be moved back out later
